Update project files in order
