+++ PetShop-Management-System.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PetShop-Management-System", "PetShop-Management-System.csproj", "{E76F8668-360F-45F6-B3A4-14D6A547EF51}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BP-ERP-Final-Project", "..\..\..\..\Downloads\BP-ERP-Final-Project\BP-ERP-Final-Project.csproj", "{94EFEC70-21C0-4790-93F1-D59D7D4EA421}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{E76F8668-360F-45F6-B3A4-14D6A547EF51}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E76F8668-360F-45F6-B3A4-14D6A547EF51}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E76F8668-360F-45F6-B3A4-14D6A547EF51}.Release|Any CPU.Build.0 = Release|Any CPU
-		{94EFEC70-21C0-4790-93F1-D59D7D4EA421}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94EFEC70-21C0-4790-93F1-D59D7D4EA421}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94EFEC70-21C0-4790-93F1-D59D7D4EA421}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94EFEC70-21C0-4790-93F1-D59D7D4EA421}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
