+++ UIs/Home.cs
@@ -10,11 +10,21 @@ using System.Windows.Forms;
 
 namespace PetShop_Management_System
 {
-    public partial class Bills : Form
+    public partial class Home : Form
     {
-        public Bills()
+        public Home()
         {
             InitializeComponent();
+        }
+
+        private void label7_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void Home_Load(object sender, EventArgs e)
+        {
+
         }
     }
 }
