+++ UIs/Loginb.resx
@@ -121,7 +121,7 @@
   <data name="pictureBox2.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAAgAAAAIACAYAAAD0eNT6AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAX
-        hgAAF4YB9EXJGQAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAACgXSURBVHhe7d1r
+        hQAAF4UBWfzodAAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAACgXSURBVHhe7d1r
         rG1rWR9wKcjBNqGcRg0UAlYrhrZa05aTphq0iVHbemlLmn6gMWmiif1EQhrQw0WQmNB4GkXQXrDSpAWj
         CYk1xbaY0GJSE6Ga9EYrYk2pWiuRUIqKIND/u1kT5h77Wbe5xhxjvO/4ffglZz9rzjHH5Xnf/5Oz91rr
         sz75yU8O5XGf9QcejkfiBfGqeHO8Pd4Z7473xQfioxfaf7da+1p7TXtte097bztGO9bD1WcBsD1tz77Y
@@ -299,7 +299,7 @@
   <data name="pictureBox1.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAAgAAAAIACAYAAAD0eNT6AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
-        vgAADr4B6kKxwAAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAChISURBVHhe7d15
+        vQAADr0BR/uQrQAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAChISURBVHhe7d15
         vG713PDx02k4zSUaVEJSmp5UGjQgGRoRQhSZop7I7UU9pNucbtzcmSoPokyVOUJJhiLNKJVSSYOKaJ5O
         Pffn+5zO3eB79tl7X+v6reH3+bxe7z/rrL2uvdb13de11m/NMDMzMzMzMzMzMzMzMzMzMzMzMzMzMzMz
         MzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzM7Mxtjieiddhf7wP+2BnrAYzMzMbSItgd/wC
