+++ UIs/Splash.resx
@@ -124,7 +124,7 @@
   <data name="pictureBox1.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAAgAAAAIACAYAAAD0eNT6AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
-        vgAADr4B6kKxwAAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAChISURBVHhe7d15
+        vQAADr0BR/uQrQAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAChISURBVHhe7d15
         vG713PDx02k4zSUaVEJSmp5UGjQgGRoRQhSZop7I7UU9pNucbtzcmSoPokyVOUJJhiLNKJVSSYOKaJ5O
         Pffn+5zO3eB79tl7X+v6reH3+bxe7z/rrL2uvdb13de11m/NMDMzMzMzMzMzMzMzMzMzMzMzMzMzMzMz
         MzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzM7Mxtjieiddhf7wP+2BnrAYzMzMbSItgd/wC
